% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBudget.R
\name{getBudget}
\alias{getBudget}
\title{getBudget}
\usage{
getBudget(df, school = "University of St Thomas (Minnesota)",
  low_pay = TRUE)
}
\arguments{
\item{df}{The dataframe of pay amount by different universities}

\item{school}{The school of the average graduate you would like to generate a budget for}

\item{low_pay}{True if the generator should use low_pay false if high_pay}
}
\value{
A tibble of the budget
}
\description{
This function gives a tibble that contains variables of different spending catagories and values based on a budget suggested by thebalance.com https://www.thebalance.com/how-to-figure-out-budget-percentages-for-money-goals-4171689
}
\examples{
getBudget(df = computerSciencePay, school = "University of St Thomas (Minnesota)", low_pay = FALSE)
}
