% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareChart.R
\name{compareChart}
\alias{compareChart}
\title{Title: compareChart}
\usage{
compareChart(school = "University of St Thomas (Minnesota)",
  whichPay = "low_pay")
}
\arguments{
\item{school}{Name}
}
\description{
Title: compareChart
}
\examples{
: getGraph(University of St Thomas (Minnesota))
}
