% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareChart.R
\name{compareChart}
\alias{compareChart}
\title{Title: CompareChart Function}
\usage{
compareChart(df, school = "University of St Thomas (Minnesota)",
  whichPay = "low_pay")
}
\arguments{
\item{whichPay}{}
}
\description{
Title: CompareChart Function
}
\examples{
: compareChart("Stanford University","low_pay")
}
