% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareChart.R
\name{compareChart}
\alias{compareChart}
\title{Title}
\usage{
compareChart(df, school = "University of St Thomas (Minnesota)",
  whichPay = "low_pay")
}
\arguments{
\item{whichPay}{}
}
\description{
Title
}
